requirements.txt
notion-client>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pywin32>=306; sys_platform == "win32"
